Move version to separate module
